backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
pandas==2.1.3
numpy==1.26.2
python-dateutil==2.8.2
gunicorn==21.2.0
